add AicController

+++ src/main/java/org/daybreak/coccinella/domain/model/Crawler.java
@@ -1,20 +1,13 @@
 package org.daybreak.coccinella.domain.model;
 
+import org.hibernate.annotations.Fetch;
+import org.hibernate.annotations.FetchMode;
 import org.springframework.http.HttpMethod;
 
 import java.io.Serializable;
 import java.util.List;
 import java.util.Set;
-import javax.persistence.CascadeType;
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-import javax.persistence.OneToMany;
-import javax.persistence.Table;
+import javax.persistence.*;
 
 /**
  * 抓取任务
@@ -28,7 +21,7 @@ public class Crawler implements Serializable {
     @Id
     @Column(name = "ID")
     @GeneratedValue
-    private Integer id;
+    private long id;
     
     @Column(name = "NAME")
     private String name;
@@ -55,11 +48,11 @@ public class Crawler implements Serializable {
     @Column(name = "METHOD")
     private HttpMethod method;
 
-    public Integer getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(long id) {
         this.id = id;
     }
